- move HasLink interface to core module
